feat: Add HomePageClient component with responsive place cards and action buttons
- Implemented HomePageClient component to display a list of third places in Charlotte.
- Integrated responsive place cards and filters for better user experience.
- Added mobile-only action buttons for random suggestions, map view, and browsing.
- Updated global CSS with new action button colors for light and dark modes.
